Move ghost code to client
It's not used by the server, so there's no reason to put it into the
shared codebase.
